Removendo valor referentes a amortizações da dívidas que foram roladas e estavam sendo contabilizadas nos estados
