+++ src/main/java/com/brasil/transparente/processor/util/constants/Constants.java
@@ -12,6 +12,13 @@ public final class Constants {
     public static final String SEM_INFORMACAO = "Sem informação";
     public static final String REPASSES = "Distribuição Constitucional ou Legal de Rec";
     public static final String AMORTIZACAO_DIVIDA = "Amortização/Refinanciamento da Dívida";
+    public static final List<String> PRINCIPAL_DIVIDA_AMORTIZACAO = List.of(
+            "PRINCIPAL DA DÍVIDA CONTRATUAL RESGATADO",
+            "Principal da Dívida Contratual Resgatado",
+            "PRINCIPAL DA DIVIDA CONTRATUAL RESGATADO",
+            "PRINCIPAL DA DIVIDA POR CONTRATO EXTERNA",
+            "AMORTIZ.DA DIVIDA INTERNA CONTRATADA"
+    );
     public static final String TRANSFERENCIA_MUNICIPIOS = "Transferencias A Municipios";
     public static final String PAGAMENTO = "Pagamento";
     public static final List<String> EXECUTIVO = List.of("PODER EXECUTIVO", "Executivo");
